Updated some config options to be read as uint16_t not uint8_t to allow for a wider param space (#229)

+++ src/lib/pipeline/LoadStoreQueue.cc
@@ -19,8 +19,9 @@ LoadStoreQueue::LoadStoreQueue(
     unsigned int maxCombinedSpace, MemoryInterface& memory,
     span<PipelineBuffer<std::shared_ptr<Instruction>>> completionSlots,
     std::function<void(span<Register>, span<RegisterValue>)> forwardOperands,
-    bool exclusive, uint8_t loadBandwidth, uint8_t storeBandwidth,
-    uint8_t permittedRequests, uint8_t permittedLoads, uint8_t permittedStores)
+    bool exclusive, uint16_t loadBandwidth, uint16_t storeBandwidth,
+    uint16_t permittedRequests, uint16_t permittedLoads,
+    uint16_t permittedStores)
     : completionSlots_(completionSlots),
       forwardOperands_(forwardOperands),
       maxCombinedSpace_(maxCombinedSpace),
@@ -38,8 +39,9 @@ LoadStoreQueue::LoadStoreQueue(
     MemoryInterface& memory,
     span<PipelineBuffer<std::shared_ptr<Instruction>>> completionSlots,
     std::function<void(span<Register>, span<RegisterValue>)> forwardOperands,
-    bool exclusive, uint8_t loadBandwidth, uint8_t storeBandwidth,
-    uint8_t permittedRequests, uint8_t permittedLoads, uint8_t permittedStores)
+    bool exclusive, uint16_t loadBandwidth, uint16_t storeBandwidth,
+    uint16_t permittedRequests, uint16_t permittedLoads,
+    uint16_t permittedStores)
     : completionSlots_(completionSlots),
       forwardOperands_(forwardOperands),
       maxLoadQueueSpace_(maxLoadQueueSpace),
@@ -374,7 +376,7 @@ void LoadStoreQueue::tick() {
   // Send memory requests adhering to set bandwidth and number of permitted
   // requests per cycle
   // Index 0: loads, index 1: stores
-  std::array<uint8_t, 2> reqCounts = {0, 0};
+  std::array<uint16_t, 2> reqCounts = {0, 0};
   std::array<uint64_t, 2> dataTransfered = {0, 0};
   std::array<bool, 2> exceededLimits = {false, false};
   auto itLoad = requestLoadQueue_.begin();
